refactor: move app fonts from the root layout to their own file
